python/requirements-core.txt
openai>=1.60.0
python-docx>=1.1.2
